// src/apis/models/FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_GetPacsCardHolderForEditOutput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderCards_PacsCardHolderCardListDto } from './FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderCards_PacsCardHolderCardListDto';
import type { FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderCompanys_PacsCardHolderCompanyListDto } from './FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderCompanys_PacsCardHolderCompanyListDto';
import type { FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderDevices_PacsCardHolderDeviceListDto } from './FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderDevices_PacsCardHolderDeviceListDto';
import type { FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderEditDto } from './FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderEditDto';
export type FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_GetPacsCardHolderForEditOutput = {
    pacsCardHolder: FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderEditDto;
    pacsCardHolderCards?: Array<FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderCards_PacsCardHolderCardListDto> | null;
    pacsCardHolderCompanys?: Array<FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderCompanys_PacsCardHolderCompanyListDto> | null;
    pacsCardHolderDevices?: Array<FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderDevices_PacsCardHolderDeviceListDto> | null;
};